Feat jwt v1

+++ src/main/resources/db/migration/V2__Create_users_table.sql
@@ -1,11 +1,11 @@
 CREATE TABLE TB_USERS (
-                          id BIGSERIAL PRIMARY KEY,
-                          name VARCHAR(255),
-                          email VARCHAR(255) NOT NULL UNIQUE,
-                          password VARCHAR(255),
-                          last_modified_date TIMESTAMP,
-                          address_id BIGINT,
-                          created_date TIMESTAMP,
-                          roles VARCHAR(255),
-                          FOREIGN KEY (address_id) REFERENCES ADDRESS(id)
+    id SERIAL PRIMARY KEY,
+    name VARCHAR(255),
+    email VARCHAR(255) NOT NULL UNIQUE,
+    password VARCHAR(255),
+    last_modified_date TIMESTAMP,
+    address_id INT,
+    created_date TIMESTAMP,
+    roles VARCHAR(255),
+    FOREIGN KEY (address_id) REFERENCES ADDRESS(id)
 );
